requirements.txt
openpyxl==3.0.10
pandas==1.4.4
Pillow==9.5.0
streamlit==1.22.0

